UserController changes and db implementation

+++ demo-rest/src/main/java/com/foodtec/demorest/service/UserService.java
@@ -2,11 +2,12 @@ package com.foodtec.demorest.service;
 
 import com.foodtec.demorest.entity.User;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 
 public interface UserService {
 
-    List<User> listUsers(String name, Date registrationDate, Double income);
-
+    List<User> listUsers(String name, Date registrationDate, BigDecimal income);
 }
